requirements.txt
flask
ldap3
python-cinderclient
python-glanceclient
python-keystoneclient
python-neutronclient
python-novaclient
python-designateclient
python-troveclient
redis
